+++ collaborative.py
@@ -1,60 +1,71 @@
-from collections import defaultdict
-
 import pandas as pd
+from sklearn.metrics.pairwise import cosine_similarity
+import numpy as np
 
-target=["Development","IT & Software"];
+# Load dataset
+ratings_df = pd.read_csv("modified_comments_with_author_id.csv")
+ratings_df.dropna(inplace=True)
 
-df=pd.read_csv("/Users/odekunleolasubomi/PycharmProjects/Recommender/Comments.csv")
+# Filter users with at least 5 comments
+user_comment_counts = ratings_df.groupby('autor_id').size()
+users_with_min_2_comments = user_comment_counts[user_comment_counts >= 5].index
+ratings_df = ratings_df[ratings_df['autor_id'].isin(users_with_min_2_comments)]
 
+# Create USER-ITEM matrix
+ratings_matrix = ratings_df.pivot(index='autor_id', columns='course_id', values='rating')
+ratings_matrix = ratings_matrix.fillna(0)  # Fill NaN with 0
 
-# df=pd.read_csv("/Users/odekunleolasubomi/PycharmProjects/Recommender/Comments.csv")
-
-
-# counter=0
-#
-# df=df.dropna(subset=['display_name'])
-# df['display_name'] = df['display_name'].str.lower().str.strip()
-#
-# display_name_set= set(df.display_name)
-#
-# display_name_dict={}
-#
-# for i in display_name_set:
-#     display_name_dict[i]=counter
-#     counter+=1
-#
-#
-# df["user_id"]= df.apply(lambda row: display_name_dict[row['display_name']], axis=1)
-# # df["user_id"] = pd.factorize(df["display_name"])[0]
-#
-# # user_counts = df['user_id'].value_counts()
-# # df = df[df['user_id'].isin(user_counts[user_counts > 1].index)]
-#
-# # Remove rows where display_name contains only one word
-# df = df[df['display_name'].str.split().str.len() > 2]
-#
-# df.to_csv("modded_comments.csv",index=False)
-#
-# # Count how many comments each user has made
-# # user_counts = df['display_name'].value_counts()
-#
-# # Add a column for the number of comments each user has made
-#
-#
-#
-# # user_counts = set(df.user_id)
-# # user_counts_dic = defaultdict(int)
-# # for i,row in df.iterrows():
-# #     user_counts_dic[row["user_id"]] += 1
-# #
-# #
-# # df["number_of_comments"]=df.apply(lambda row : user_counts_dic[row["user_id"]], axis=1)
-#
-# user_counts_dic = df.groupby("user_id").size().to_dict()
-# df["number_of_comments"] = df["user_id"].map(user_counts_dic)
-#
-# df.to_csv("avdd.csv",index=False)
+# Compute user-user similarity (cosine similarity)
+user_similarity = cosine_similarity(ratings_matrix)
+user_similarity_df = pd.DataFrame(user_similarity, index=ratings_matrix.index, columns=ratings_matrix.index)
 
 
+# Function to predict rating for a user on a given course
+def predict_rating(user_id, course_id):
+    if user_id not in ratings_matrix.index or course_id not in ratings_matrix.columns:
+        return None  # User or course not found
+
+    # Find similar users
+    similar_users = user_similarity_df[user_id].drop(user_id).sort_values(ascending=False)
+
+    # Ratings given by similar users for the course
+    user_ratings = ratings_matrix.loc[similar_users.index, course_id]
+
+    # Filter users who actually rated this course
+    valid_ratings = user_ratings[user_ratings > 0]
+
+    if valid_ratings.empty:
+        return None  # No similar users have rated this course
+
+    # Compute weighted rating prediction (weighted by similarity)
+    weights = similar_users.loc[valid_ratings.index]
+    if weights.sum() == 0:  # Avoid division by zero
+        return None
+    predicted_rating = np.dot(valid_ratings, weights) / weights.sum()
+
+    return predicted_rating
 
 
+# Function to get recommendations for a user
+def get_recommendations(user_id, threshold=2.5, top_n=5):
+    if user_id not in ratings_matrix.index:
+        print(f"User ID {user_id} not found.")
+        return None
+
+    predictions = {}
+
+    for course_id in ratings_matrix.columns:
+        if ratings_matrix.loc[user_id, course_id] == 0:  # If user hasn't rated it
+            predicted_rating = predict_rating(user_id, course_id)
+            if predicted_rating is not None and predicted_rating > threshold:
+                predictions[course_id] = predicted_rating
+
+    # Sort recommendations by predicted rating
+    recommended_courses = sorted(predictions.items(), key=lambda x: x[1], reverse=True)[:top_n]
+
+    return recommended_courses
+
+
+# Example: Get top 5 recommendations for user 1001
+recommended_courses_user_1 = get_recommendations(23448456, threshold=2.5, top_n=5)
+print("Recommended courses for User 1001:", recommended_courses_user_1)
